feat(TDP-1691): remove react native from article label
