+++ packages/article-label/__tests__/web/jest.config.js
@@ -1,6 +1,3 @@
 const jestConfigurator = require("@times-components/jest-configurator").default;
-const path = require("path");
 
-module.exports = jestConfigurator("web", __dirname, {
-  setupFilesAfterEnv: path.join(__dirname, "./serializers")
-});
+module.exports = jestConfigurator("web", __dirname);
