--- packages/article-label/__tests__/web/serializers.js
@@ -1,4 +1,0 @@
-const { AppRegistry } = require("react-native-web");
-const { addSerializers, rnw } = require("@times-components/jest-serializer");
-
-addSerializers(expect, rnw(AppRegistry));
