+++ packages/article-label/src/article-label.js
@@ -1,13 +1,16 @@
 import React from "react";
-import { Text } from "react-native";
-import { gqlRgbaToHex } from "@times-components/utils";
+import { checkStylesForUnits, gqlRgbaToHex, TcText } from "@times-components/utils";
 import PropTypes from "prop-types";
 import styles from "./style";
 
+const articleStyle = (color) => {
+  return {...checkStylesForUnits(styles.title), color: gqlRgbaToHex(color) || color};
+};
+
 const ArticleLabel = ({ color, title }) => (
-  <Text style={[styles.title, { color: gqlRgbaToHex(color) || color }]}>
+  <TcText style={articleStyle(color)}>
     {title.toUpperCase()}
-  </Text>
+  </TcText>
 );
 
 ArticleLabel.propTypes = {
