+++ packages/article-label/src/style/index.js
@@ -1,13 +1,12 @@
-import { StyleSheet } from "react-native";
 import sharedStyles from "./shared";
 
-const styles = StyleSheet.create({
+const styles = {
   ...sharedStyles,
   title: {
     ...sharedStyles.title,
     lineHeight: 12,
     letterSpacing: 1
   }
-});
+};
 
 export default styles;
