+++ packages/responsive/package.json
@@ -32,7 +32,7 @@
   },
   "homepage": "https://github.com/newsuk/times-components#readme",
   "dependencies": {
-    "@times-components/styleguide": "3.38.37",
+    "@times-components/styleguide": "3.38.38",
     "@times-components/utils": "6.8.0",
     "prop-types": "15.7.2"
   },
